3 Cliff Place
"On June 3rd, that is, on Monday last, McCarthy left his house at Hatherley about three in the
afternoon and walked down to the Boscombe Pool, which is a small lake formed by the spreading out of
the stream which runs down the Boscombe Valley. He had been out with his serving-man in the morning
at Ross, and he had told the man that he must hurry, as he had an appointment of importance to keep
at three. From that appointment he never came back alive.

"From Hatherley Farm-house to the Boscombe Pool is a quarter of a mile, and two people saw him as he
passed over this ground. One was an old woman, whose name is not mentioned, and the other was
William Crowder, a game-keeper in the employ of Mr. Turner. Both these witnesses depose that Mr.
McCarthy was walking alone. The game-keeper adds that within a few minutes of his seeing Mr.
McCarthy pass he had seen his son, Mr. James McCarthy, going the same way with a gun under his arm.
To the best of his belief, the father was actually in sight at the time, and the son was following
him. He thought no more of the matter until he heard in the evening of the tragedy that had
occurred.

"The two McCarthys were seen after the time when William Crowder, the game-keeper, lost sight of
them. The Boscombe Pool is thickly wooded round, with just a fringe of grass and of reeds round the
edge. A girl of fourteen, Patience Moran, who is the daughter of the lodge-keeper of the Boscombe
Valley estate, was in one of the woods picking flowers. She states that while she was there she saw,
at the border of the wood and close by the lake, Mr. McCarthy and his son, and that they appeared to
be having a violent quarrel. She heard Mr. McCarthy the elder using very strong language to his son,
and she saw the latter raise up his hand as if to strike his father. She was so frightened by their
violence that she ran away and told her mother when she reached home that she had left the two
McCarthys quarrelling near Boscombe Pool, and that she was afraid that they were going to fight. She
had hardly said the words when young Mr. McCarthy came running up to the lodge to say that he had
found his father dead in the wood, and to ask for the help of the lodge-keeper. He was much excited,
without either his gun or his hat, and his right hand and sleeve were observed to be stained with
fresh blood. On following him they found the dead body stretched out upon the grass beside the pool.
The head had been beaten in by repeated blows of some heavy and blunt weapon. The injuries were such
as might very well have been inflicted by the butt-end of his son's gun, which was found lying on
the grass within a few paces of the body. Under these circumstances the young man was instantly
arrested, and a verdict of 'wilful murder' having been returned at the inquest on Tuesday, he was on
Wednesday brought before the magistrates at Ross, who have referred the case to the next Assizes.
Those are the main facts of the case as they came out before the coroner and the police-court."