96 Valentia Road
"With the result of driving it through the ventilator."

"And also with the result of causing it to turn upon its master at the other side. Some of the blows
of my cane came home and roused its snakish temper, so that it flew upon the first person it saw. In
this way I am no doubt indirectly responsible for Dr. Grimesby Roylott's death, and I cannot say
that it is likely to weigh very heavily upon my conscience."

Of all the problems which have been submitted to my friend, Mr. Sherlock Holmes, for solution during
the years of our intimacy, there were only two which I was the means of introducing to his notice--
that of Mr. Hatherley's thumb, and that of Colonel Warburton's madness. Of these the latter may have
afforded a finer field for an acute and original observer, but the other was so strange in its
inception and so dramatic in its details that it may be the more worthy of being placed upon record,
even if it gave my friend fewer openings for those deductive methods of reasoning by which he
achieved such remarkable results. The story has, I believe, been told more than once in the
newspapers, but, like all such narratives, its effect is much less striking when set forth en bloc
in a single half-column of print than when the facts slowly evolve before your own eyes, and the
mystery clears gradually away as each new discovery furnishes a step which leads on to the complete
truth. At the time the circumstances made a deep impression upon me, and the lapse of two years has
hardly served to weaken the effect.

It was in the summer of '89, not long after my marriage, that the events occurred which I am now
about to summarise. I had returned to civil practice and had finally abandoned Holmes in his Baker
Street rooms, although I continually visited him and occasionally even persuaded him to forgo his
Bohemian habits so far as to come and visit us. My practice had steadily increased, and as I
happened to live at no very great distance from Paddington Station, I got a few patients from among
the officials. One of these, whom I had cured of a painful and lingering disease, was never weary of
advertising my virtues and of endeavouring to send me on every sufferer over whom he might have any
influence.

One morning, at a little before seven o'clock, I was awakened by the maid tapping at the door to
announce that two men had come from Paddington and were waiting in the consulting-room. I dressed
hurriedly, for I knew by experience that railway cases were seldom trivial, and hastened downstairs.
As I descended, my old ally, the guard, came out of the room and closed the door tightly behind him.

"I've got him here," he whispered, jerking his thumb over his shoulder; "he's all right."