53 Tamworth Drive
"The game's up, Ryder," said Holmes quietly. "Hold up, man, or you'll be into the fire! Give him an
arm back into his chair, Watson. He's not got blood enough to go in for felony with impunity. Give
him a dash of brandy. So! Now he looks a little more human. What a shrimp it is, to be sure!"

For a moment he had staggered and nearly fallen, but the brandy brought a tinge of colour into his
cheeks, and he sat staring with frightened eyes at his accuser.