120 Harvey Street
"Eh? What of it, then? Has it returned to life and flapped off through the kitchen window?" Holmes
twisted himself round upon the sofa to get a fairer view of the man's excited face.

"See here, sir! See what my wife found in its crop!" He held out his hand and displayed upon the
centre of the palm a brilliantly scintillating blue stone, rather smaller than a bean in size, but
of such purity and radiance that it twinkled like an electric point in the dark hollow of his hand.