78 Wantage Close
"Yes, sir; for he had started to England just before it arrived."

"Ha! that was unfortunate. Your wedding was arranged, then, for the Friday. Was it to be in church?"

"Yes, sir, but very quietly. It was to be at St. Saviour's, near King's Cross, and we were to have
breakfast afterwards at the St. Pancras Hotel. Hosmer came for us in a hansom, but as there were two
of us he put us both into it and stepped himself into a four-wheeler, which happened to be the only
other cab in the street. We got to the church first, and when the four-wheeler drove up we waited
for him to step out, but he never did, and when the cabman got down from the box and looked there
was no one there! The cabman said that he could not imagine what had become of him, for he had seen
him get in with his own eyes. That was last Friday, Mr. Holmes, and I have never seen or heard
anything since then to throw any light upon what became of him."

"It seems to me that you have been very shamefully treated," said Holmes.

"Oh, no, sir! He was too good and kind to leave me so. Why, all the morning he was saying to me
that, whatever happened, I was to be true; and that even if something quite unforeseen occurred to
separate us, I was always to remember that I was pledged to him, and that he would claim his pledge
sooner or later. It seemed strange talk for a wedding-morning, but what has happened since gives a
meaning to it."