128 Beaconside
"I am in hopes that she does."

"And why in hopes?"

"Because it would spare your Majesty all fear of future annoyance. If the lady loves her husband,
she does not love your Majesty. If she does not love your Majesty, there is no reason why she should
interfere with your Majesty's plan."

"It is true. And yet--Well! I wish she had been of my own station! What a queen she would have
made!" He relapsed into a moody silence, which was not broken until we drew up in Serpentine Avenue.

The door of Briony Lodge was open, and an elderly woman stood upon the steps. She watched us with a
sardonic eye as we stepped from the brougham.