173 Beaconside
"It was lost, if I remember aright, at the Hotel Cosmopolitan," I remarked.

"Precisely so, on December 22nd, just five days ago. John Horner, a plumber, was accused of having
abstracted it from the lady's jewel-case. The evidence against him was so strong that the case has
been referred to the Assizes. I have some account of the matter here, I believe." He rummaged amid
his newspapers, glancing over the dates, until at last he smoothed one out, doubled it over, and
read the following paragraph:

"Hotel Cosmopolitan Jewel Robbery. John Horner, 26, plumber, was brought up upon the charge of
having upon the 22nd inst. abstracted from the jewel-case of the Countess of Morcar the valuable gem
known as the blue carbuncle. James Ryder, upper-attendant at the hotel, gave his evidence to the
effect that he had shown Horner up to the dressing-room of the Countess of Morcar upon the day of
the robbery in order that he might solder the second bar of the grate, which was loose. He had
remained with Horner some little time, but had finally been called away. On returning, he found that
Horner had disappeared, that the bureau had been forced open, and that the small morocco casket in
which, as it afterwards transpired, the Countess was accustomed to keep her jewel, was lying empty
upon the dressing-table. Ryder instantly gave the alarm, and Horner was arrested the same evening;
but the stone could not be found either upon his person or in his rooms. Catherine Cusack, maid to
the Countess, deposed to having heard Ryder's cry of dismay on discovering the robbery, and to
having rushed into the room, where she found matters as described by the last witness. Inspector
Bradstreet, B division, gave evidence as to the arrest of Horner, who struggled frantically, and
protested his innocence in the strongest terms. Evidence of a previous conviction for robbery having
been given against the prisoner, the magistrate refused to deal summarily with the offence, but
referred it to the Assizes. Horner, who had shown signs of intense emotion during the proceedings,
fainted away at the conclusion and was carried out of court."

"Hum! So much for the police-court," said Holmes thoughtfully, tossing aside the paper. "The
question for us now to solve is the sequence of events leading from a rifled jewel-case at one end
to the crop of a goose in Tottenham Court Road at the other. You see, Watson, our little deductions
have suddenly assumed a much more important and less innocent aspect. Here is the stone; the stone
came from the goose, and the goose came from Mr. Henry Baker, the gentleman with the bad hat and all
the other characteristics with which I have bored you. So now we must set ourselves very seriously
to finding this gentleman and ascertaining what part he has played in this little mystery. To do
this, we must try the simplest means first, and these lie undoubtedly in an advertisement in all the
evening papers. If this fail, I shall have recourse to other methods."

"What will you say?"