21 Beaconside
"This is a very deep business," he said at last. "There are a thousand details which I should desire
to know before I decide upon our course of action. Yet we have not a moment to lose. If we were to
come to Stoke Moran to-day, would it be possible for us to see over these rooms without the
knowledge of your stepfather?"

"As it happens, he spoke of coming into town to-day upon some most important business. It is
probable that he will be away all day, and that there would be nothing to disturb you. We have a
housekeeper now, but she is old and foolish, and I could easily get her out of the way."

"Excellent. You are not averse to this trip, Watson?"

"By no means."

"Then we shall both come. What are you going to do yourself?"

"I have one or two things which I would wish to do now that I am in town. But I shall return by the
twelve o'clock train, so as to be there in time for your coming."