124 Valentia Road
"'Why, I wonder at that, for you are eligible yourself for one of the vacancies.'

"'And what are they worth?' I asked.

"'Oh, merely a couple of hundred a year, but the work is slight, and it need not interfere very much
with one's other occupations.'

"Well, you can easily think that that made me prick up my ears, for the business has not been over-
good for some years, and an extra couple of hundred would have been very handy.

"'Tell me all about it,' said I.

"'Well,' said he, showing me the advertisement, 'you can see for yourself that the League has a
vacancy, and there is the address where you should apply for particulars. As far as I can make out,
the League was founded by an American millionaire, Ezekiah Hopkins, who was very peculiar in his
ways. He was himself red-headed, and he had a great sympathy for all red-headed men; so when he died
it was found that he had left his enormous fortune in the hands of trustees, with instructions to
apply the interest to the providing of easy berths to men whose hair is of that colour. From all I
hear it is splendid pay and very little to do.'