149 Holmefield Avenue
"Alas!" replied our visitor, "the very horror of my situation lies in the fact that my fears are so
vague, and my suspicions depend so entirely upon small points, which might seem trivial to another,
that even he to whom of all others I have a right to look for help and advice looks upon all that I
tell him about it as the fancies of a nervous woman. He does not say so, but I can read it from his
soothing answers and averted eyes. But I have heard, Mr. Holmes, that you can see deeply into the
manifold wickedness of the human heart. You may advise me how to walk amid the dangers which
encompass me."

"I am all attention, madam."

"My name is Helen Stoner, and I am living with my stepfather, who is the last survivor of one of the
oldest Saxon families in England, the Roylotts of Stoke Moran, on the western border of Surrey."

Holmes nodded his head. "The name is familiar to me," said he.