28 Holmefield Avenue
"On the contrary, this is the right side."

"The right side? You're mad! Here is the note written in pencil over here."

"And over here is what appears to be the fragment of a hotel bill, which interests me deeply."

"There's nothing in it. I looked at it before," said Lestrade.

"'Oct. 4th, rooms 8s., breakfast 2s. 6d., cocktail 1s., lunch 2s. 6d., glass sherry, 8d.' I see
nothing in that."