94 Larch Walk
"Yes, it would have been of no use to anyone had we not done so. But I presume that this other goose
upon the sideboard, which is about the same weight and perfectly fresh, will answer your purpose
equally well?"

"Oh, certainly, certainly," answered Mr. Baker with a sigh of relief.

"Of course, we still have the feathers, legs, crop, and so on of your own bird, so if you wish--"

The man burst into a hearty laugh. "They might be useful to me as relics of my adventure," said he,
"but beyond that I can hardly see what use the disjecta membra of my late acquaintance are going to
be to me. No, sir, I think that, with your permission, I will confine my attentions to the excellent
bird which I perceive upon the sideboard."

Sherlock Holmes glanced sharply across at me with a slight shrug of his shoulders.