3 Valentia Road
"I did as I was told, and at the same instant Mrs. Rucastle drew down the blind. That was a week
ago, and from that time I have not sat again in the window, nor have I worn the blue dress, nor seen
the man in the road."

"Pray continue," said Holmes. "Your narrative promises to be a most interesting one."

"You will find it rather disconnected, I fear, and there may prove to be little relation between the
different incidents of which I speak. On the very first day that I was at the Copper Beeches, Mr.
Rucastle took me to a small outhouse which stands near the kitchen door. As we approached it I heard
the sharp rattling of a chain, and the sound as of a large animal moving about.

"'Look in here!' said Mr. Rucastle, showing me a slit between two planks. 'Is he not a beauty?'

"I looked through and was conscious of two glowing eyes, and of a vague figure huddled up in the
darkness.

"'Don't be frightened,' said my employer, laughing at the start which I had given. 'It's only Carlo,
my mastiff. I call him mine, but really old Toller, my groom, is the only man who can do anything
with him. We feed him once a day, and not too much then, so that he is always as keen as mustard.
Toller lets him loose every night, and God help the trespasser whom he lays his fangs upon. For
goodness' sake don't you ever on any pretext set your foot over the threshold at night, for it's as
much as your life is worth.'