113 Corbet Close
"'Tell me all about it,' said I.

"'Well,' said he, showing me the advertisement, 'you can see for yourself that the League has a
vacancy, and there is the address where you should apply for particulars. As far as I can make out,
the League was founded by an American millionaire, Ezekiah Hopkins, who was very peculiar in his
ways. He was himself red-headed, and he had a great sympathy for all red-headed men; so when he died
it was found that he had left his enormous fortune in the hands of trustees, with instructions to
apply the interest to the providing of easy berths to men whose hair is of that colour. From all I
hear it is splendid pay and very little to do.'

"'But,' said I, 'there would be millions of red-headed men who would apply.'

"'Not so many as you might think,' he answered. 'You see it is really confined to Londoners, and to
grown men. This American had started from London when he was young, and he wanted to do the old town
a good turn. Then, again, I have heard it is no use your applying if your hair is light red, or dark
red, or anything but real bright, blazing, fiery red. Now, if you cared to apply, Mr. Wilson, you
would just walk in; but perhaps it would hardly be worth your while to put yourself out of the way
for the sake of a few hundred pounds.'

"Now, it is a fact, gentlemen, as you may see for yourselves, that my hair is of a very full and
rich tint, so that it seemed to me that if there was to be any competition in the matter I stood as
good a chance as any man that I had ever met. Vincent Spaulding seemed to know so much about it that
I thought he might prove useful, so I just ordered him to put up the shutters for the day and to
come right away with me. He was very willing to have a holiday, so we shut the business up and
started off for the address that was given us in the advertisement.

"I never hope to see such a sight as that again, Mr. Holmes. From north, south, east, and west every
man who had a shade of red in his hair had tramped into the city to answer the advertisement. Fleet
Street was choked with red-headed folk, and Pope's Court looked like a coster's orange barrow. I
should not have thought there were so many in the whole country as were brought together by that
single advertisement. Every shade of colour they were--straw, lemon, orange, brick, Irish-setter,
liver, clay; but, as Spaulding said, there were not many who had the real vivid flame-coloured tint.
When I saw how many were waiting, I would have given it up in despair; but Spaulding would not hear
of it. How he did it I could not imagine, but he pushed and pulled and butted until he got me
through the crowd, and right up to the steps which led to the office. There was a double stream upon
the stair, some going up in hope, and some coming back dejected; but we wedged in as well as we
could and soon found ourselves in the office."