9 Valentia Road
"In the first place, both my friend and I must spend the night in your room."

Both Miss Stoner and I gazed at him in astonishment.

"Yes, it must be so. Let me explain. I believe that that is the village inn over there?"

"Yes, that is the Crown."

"Very good. Your windows would be visible from there?"

"Certainly."