111 Larch Walk
What a time it seemed! From comparing notes afterwards it was but an hour and a quarter, yet it
appeared to me that the night must have almost gone and the dawn be breaking above us. My limbs were
weary and stiff, for I feared to change my position; yet my nerves were worked up to the highest
pitch of tension, and my hearing was so acute that I could not only hear the gentle breathing of my
companions, but I could distinguish the deeper, heavier in-breath of the bulky Jones from the thin,
sighing note of the bank director. From my position I could look over the case in the direction of
the floor. Suddenly my eyes caught the glint of a light.

At first it was but a lurid spark upon the stone pavement. Then it lengthened out until it became a
yellow line, and then, without any warning or sound, a gash seemed to open and a hand appeared, a
white, almost womanly hand, which felt about in the centre of the little area of light. For a minute
or more the hand, with its writhing fingers, protruded out of the floor. Then it was withdrawn as
suddenly as it appeared, and all was dark again save the single lurid spark which marked a chink
between the stones.

Its disappearance, however, was but momentary. With a rending, tearing sound, one of the broad,
white stones turned over upon its side and left a square, gaping hole, through which streamed the
light of a lantern. Over the edge there peeped a clean-cut, boyish face, which looked keenly about
it, and then, with a hand on either side of the aperture, drew itself shoulder-high and waist-high,
until one knee rested upon the edge. In another instant he stood at the side of the hole and was
hauling after him a companion, lithe and small like himself, with a pale face and a shock of very
red hair.