30 Larch Walk
"'I beg that you will state your business, sir,' said I; 'my time is of value.' Heaven forgive me
for that last sentence, but the words came to my lips.

"'How would fifty guineas for a night's work suit you?' he asked.

"'Most admirably.'

"'I say a night's work, but an hour's would be nearer the mark. I simply want your opinion about a
hydraulic stamping machine which has got out of gear. If you show us what is wrong we shall soon set
it right ourselves. What do you think of such a commission as that?'

"'The work appears to be light and the pay munificent.'

"'Precisely so. We shall want you to come to-night by the last train.'