19 Glendale
"But you would not advise me to refuse?"

"I confess that it is not the situation which I should like to see a sister of mine apply for."

"What is the meaning of it all, Mr. Holmes?"

"Ah, I have no data. I cannot tell. Perhaps you have yourself formed some opinion?"