121 Holmefield Avenue
"And what did you see?"

"What I expected to see."

"Why did you beat the pavement?"

"My dear doctor, this is a time for observation, not for talk. We are spies in an enemy's country.
We know something of Saxe-Coburg Square. Let us now explore the parts which lie behind it."