53 Sunningdale Drive
"That was it," said Holmes, nodding approvingly; "I have no doubt of it. But have you never been
prosecuted for begging?"

"Many times; but what was a fine to me?"

"It must stop here, however," said Bradstreet. "If the police are to hush this thing up, there must
be no more of Hugh Boone."

"I have sworn it by the most solemn oaths which a man can take."