64 Beaconside
"What of this 'Cooee!' then?"

"Well, obviously it could not have been meant for the son. The son, as far as he knew, was in
Bristol. It was mere chance that he was within earshot. The 'Cooee!' was meant to attract the
attention of whoever it was that he had the appointment with. But 'Cooee' is a distinctly Australian
cry, and one which is used between Australians. There is a strong presumption that the person whom
McCarthy expected to meet him at Boscombe Pool was someone who had been in Australia."

"What of the rat, then?"