2 Tamworth Drive
"Mrs. Rucastle seemed to me to be colourless in mind as well as in feature. She impressed me neither
favourably nor the reverse. She was a nonentity. It was easy to see that she was passionately
devoted both to her husband and to her little son. Her light grey eyes wandered continually from one
to the other, noting every little want and forestalling it if possible. He was kind to her also in
his bluff, boisterous fashion, and on the whole they seemed to be a happy couple. And yet she had
some secret sorrow, this woman. She would often be lost in deep thought, with the saddest look upon
her face. More than once I have surprised her in tears. I have thought sometimes that it was the
disposition of her child which weighed upon her mind, for I have never met so utterly spoiled and so
ill-natured a little creature. He is small for his age, with a head which is quite
disproportionately large. His whole life appears to be spent in an alternation between savage fits
of passion and gloomy intervals of sulking. Giving pain to any creature weaker than himself seems to
be his one idea of amusement, and he shows quite remarkable talent in planning the capture of mice,
little birds, and insects. But I would rather not talk about the creature, Mr. Holmes, and, indeed,
he has little to do with my story."

"I am glad of all details," remarked my friend, "whether they seem to you to be relevant or not."