82 Valentia Road
The telegram which we eventually received came late one night just as I was thinking of turning in
and Holmes was settling down to one of those all-night chemical researches which he frequently
indulged in, when I would leave him stooping over a retort and a test-tube at night and find him in
the same position when I came down to breakfast in the morning. He opened the yellow envelope, and
then, glancing at the message, threw it across to me.

"Just look up the trains in Bradshaw," said he, and turned back to his chemical studies.

The summons was a brief and urgent one.

Please be at the Black Swan Hotel at Winchester at midday to-morrow [it said]. Do come! I am at my
wit's end. Hunter.

"Will you come with me?" asked Holmes, glancing up.

"I should wish to."