84 Valentia Road
"Terrible! She is even more affected than I."

"You have neither of you any doubt as to your son's guilt?"

"How can we have when I saw him with my own eyes with the coronet in his hands."

"I hardly consider that a conclusive proof. Was the remainder of the coronet at all injured?"

"Yes, it was twisted."

"Do you not think, then, that he might have been trying to straighten it?"