143 Holmefield Avenue
"That also I could fathom."

"Then they carried me in. She was bound to have me in. What else could she do? And into her sitting-
room, which was the very room which I suspected. It lay between that and her bedroom, and I was
determined to see which. They laid me on a couch, I motioned for air, they were compelled to open
the window, and you had your chance."